Order checkout

+++ src/main/kotlin/mimsoft/io/features/order/price/OrderPriceDto.kt
@@ -3,12 +3,12 @@ package mimsoft.io.features.order.price
 data class OrderPriceDto(
     val id: Long? = null,
     val orderId: Long? = null,
-    val deliveryPrice: Long? = null,
-    val deliveryDiscount: Long? = null,
-    val deliveryPromo: Long? = null,
-    val productPrice: Long? = null,
-    val productDiscount: Long? = null,
-    val productPromo: Long? = null,
-    val totalPrice: Long? = null,
-    val totalDiscount: Long? = null
+    val deliveryPrice: Double? = null,
+    val deliveryDiscount: Double? = null,
+    val deliveryPromo: Double? = null,
+    val productPrice: Double? = null,
+    val productDiscount: Double? = null,
+    val productPromo: Double? = null,
+    val totalPrice: Double? = null,
+    val totalDiscount: Double? = null
 )
